% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{F_CalPMMT}
\alias{F_CalPMMT}
\title{F_CalPMMT}
\usage{
F_CalPMMT(mz, MRP = 17500, RefMZ = 200)
}
\description{
Calculate the peak matching mass tolerance (PMMT)
according to the mass resolving power (MRP) and reference m/z (RefMZ)
}
