% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{F_CountSpectraPerInchikey}
\alias{F_CountSpectraPerInchikey}
\title{F_CountSpectraPerInchikey}
\usage{
F_CountSpectraPerInchikey(Inchikey)
}
\description{
Count the number of spectra for each inchikey (first 14 charactors)
}
