% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralMatching.R
\name{F_DynamicMatching}
\alias{F_DynamicMatching}
\title{F_DynamicMatching}
\usage{
F_DynamicMatching(top = top, bottom = bottom, RefMZ = 200, MRP = 17500)
}
\description{
Matching between query and library using the dynamic tolerance
}
