% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_FDRCutoff}
\alias{F_FDRCutoff}
\title{F_FDRCutoff}
\usage{
F_FDRCutoff(DT, FDRCutoff, FileName, qvalueCutoff = 0.35)
}
\description{
Combined functions with the following goals
\enumerate{
\item Calculate the PEP score based on the target decoy distribution
\item Calculate the estimated qValue based on PEP accumulation
\item Calculate the estimated qValue based on target decoy
\item Calculate the actual qValue
\item Calculate the identification rate
\item Density plot for target and decoy scores
\item Thres .VS. Qvalue
\item Estimated q-value .VS. Identification Rate
\item True q-value .VS. Estimated q-value
\item Bland-altman plot for True q-value - Estimated q-value
}
}
