% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{F_getPEPFromScoreLambda}
\alias{F_getPEPFromScoreLambda}
\title{F_getPEPFromScoreLambda}
\usage{
F_getPEPFromScoreLambda(targetScores, decoyScores, FigName)
}
\description{
This function is modified based on the qvality.py file in \link{triqler} (https://github.com/statisticalbiotechnology/triqler/blob/master/triqler/qvality.py)
Get PEP score according to the target scores and decoy scores, a regression figure is exported as well
}
