% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_gghistogram}
\alias{F_gghistogram}
\title{F_gghistogram}
\usage{
F_gghistogram(z, name, BinNum = 30)
}
\description{
histograms for the target and decoy scores at different peak matching
mass tolerance of dynamic, 0.005 Da, 0.028 Da, 0.050 Da
}
