% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_plot.roc}
\alias{F_plot.roc}
\title{F_plot.roc}
\usage{
F_plot.roc(z, name, BinNum = 30)
}
\description{
ROC plot for scores at different peak matching
mass tolerance of dynamic, 0.005 Da, 0.028 Da, 0.050 Da
}
