% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_PlotcombinePeaks}
\alias{F_PlotcombinePeaks}
\title{F_PlotcombinePeaks}
\usage{
F_PlotcombinePeaks(MetaL, inchi, Single, Sensus, Color)
}
\description{
Consensus plot of the spectra before and after combining
}
