% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_PLotMirOfftarget}
\alias{F_PLotMirOfftarget}
\title{F_PLotMirOfftarget}
\usage{
F_PLotMirOfftarget(DT, MSMS)
}
\description{
Plot to show different compounds with similar MS2 spectra pattern
This function needs the another function of F_PLotMir
}
