% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_PlotMSMSOriginal}
\alias{F_PlotMSMSOriginal}
\title{F_PlotMSMSOriginal}
\usage{
F_PlotMSMSOriginal(alignment, top_plot, bottom_plot, Lwd = 1, label, dp)
}
\description{
Original mirror plot to show the matched and unmatched peaks between query and library
this function is simple compared to the F_PlotMSMS function
}
