% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_SpectrumLoop}
\alias{F_SpectrumLoop}
\title{F_SpectrumLoop}
\usage{
F_SpectrumLoop(Spectra, FT, mztol)
}
\description{
In a loop, calcualte the mirror plots between query and library with
fixed mztol of 0.005 Da and dynamic mztol
This function needs F_SpectrumSingle and F_SpectrumSimilarity
}
