% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_SpectrumSimilarity}
\alias{F_SpectrumSimilarity}
\title{F_SpectrumSimilarity}
\usage{
F_SpectrumSimilarity(
  spec.top,
  spec.bottom,
  mztol = 0.005,
  mz,
  rt,
  int,
  ft,
  b = 0,
  x.threshold = 0
)
}
\description{
Plot the mirror plot and calculate the similarity score between query and library with
fixed mztol of 0.005 Da and dynamic mztol
}
