% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{F_SpectrumSingle}
\alias{F_SpectrumSingle}
\title{F_SpectrumSingle}
\usage{
F_SpectrumSingle(x)
}
\description{
In a single, calcualte the mirror plots between query and library with
fixed mztol of 0.005 Da and dynamic mztol
This function needs F_SpectrumSimilarity
}
